pyproject.toml
[build-system]
requires = [
    "mypy[mypyc]~=1.11",
    "numpy~=2.0",  # because of the mypy plugin
    "setuptools~=73.0",
]
build-backend = "setuptools.build_meta"

[project]
name = "video-tools"
version = "0.0.1"
authors = [
    { name="James Johnston", email="johnstonj.public@codenest.com" },
]
description = "A small collection of utilities for working with video preservation."
readme = "README.md"
license = {file = "LICENSE"}
requires-python = ">=3.12"
dependencies = [
    "av~=12.3",
    "colorama~=0.4.6",
    "numpy~=2.0",
    "PyYAML~=6.0",
    "scikit-video~=1.1",
]

[project.optional-dependencies]
dev = [
    "mypy[mypyc]~=1.11",
    "pytest~=8.3",
    "pytest-cov~=5.0",
    "types-colorama~=0.4.15",
    "types-pyyaml~=6.0",
]

[project.scripts]
analyze_virtualdub_timing_log = "video_tools.analyze_virtualdub_timing_log:main"
dv_dif_dump = "video_tools.dv_dif_dump:main"
dv_merge = "video_tools.dv_merge:main"
dv_repair = "video_tools.dv_repair:main"
dv_resample_audio = "video_tools.dv_resample_audio:main"
top_line_errors = "video_tools.top_line_errors:main"

[tool.mypy]
plugins = [
    "numpy.typing.mypy_plugin"
]
files = "src/**/*.py,tests/**/*.py"
strict = true
# These are additional checks not covered by strict:
disallow_any_unimported = true
enable_error_code = """\
    ignore-without-code,\
    truthy-bool,\
    truthy-iterable,\
    redundant-expr,\
    redundant-self,\
    unused-awaitable\
    """
warn_unreachable = true

[tool.ruff]
line-length = 100

[tool.ruff.lint]
extend-select = ["E501", "I", "RUF"]

[tool.setuptools.packages.find]
where = ["src"]
